(refactor) create component/container folders
