--- client/components/footer.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-export default class Footer extends Component {
-
-  render() {
-    return <div id="footer"><h1>Footer</h1></div>;
-  }
-
-};
