--- client/components/rightNav.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-export default class RightNav extends Component {
-
-  render() {
-    return <div id="right"><h1>Right Panel</h1></div>;
-  }
-
-};
