+++ client/containers/rightNav.jsx
@@ -1,0 +1,16 @@
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
+
+export default class RightNav extends Component {
+
+  render() {
+    return <div id="right"><h1>Right Panel</h1></div>;
+  }
+
+};
+
+const select = (state) => {
+  return {};
+};
+
+export default connect(select)(RightNav);
